+++ bcbio/variation/coverage.py
@@ -2,20 +2,18 @@
 
 Provides estimates of coverage intervals based on callable regions
 """
+import csv
 import itertools
 import os
 import shutil
-import subprocess
 import yaml
 
 import pybedtools
-import pandas as pd
 import numpy as np
 import pysam
 
 from bcbio.variation.bedutils import clean_file
-from bcbio.utils import (file_exists, chdir, safe_makedir,
-                         append_stem, copy_plus)
+from bcbio.utils import (file_exists, append_stem, copy_plus)
 from bcbio import bam, utils
 from bcbio.bam import ref, sambamba
 from bcbio.distributed.transaction import file_transaction
@@ -143,14 +141,16 @@ def _write_cache(cache, cache_file):
     with open(cache_file, "w") as out_handle:
         yaml.safe_dump(cache, out_handle, default_flow_style=False, allow_unicode=False)
 
-def get_average_coverage(data, bam_file, bed_file=None, target_name="genome"):
+def get_average_coverage(target_name, bed_file, data, bam_file=None):
+    if not bam_file:
+        bam_file = dd.get_align_bam(data) or dd.get_work_bam(data)
     cache_file = _get_cache_file(data, target_name)
     cache = _read_cache(cache_file, [bam_file, bed_file])
     if "avg_coverage" in cache:
         return int(cache["avg_coverage"])
 
     if bed_file:
-        avg_cov = _average_bed_coverage(data, bed_file, bam_file, target_name=target_name)
+        avg_cov = _average_bed_coverage(bed_file, data)
     else:
         avg_cov = _average_genome_coverage(data, bam_file)
 
@@ -170,165 +170,75 @@ def _average_genome_coverage(data, bam_file):
     avg_cov = float(read_counts * read_size) / total
     return avg_cov
 
-def _average_bed_coverage(data, bed_file, bam_file, target_name):
-    sambamba_depth_file = regions_coverage(data, bed_file, bam_file, target_name)
+def _average_bed_coverage(bed_file, data):
+    sambamba_depth_file = regions_coverage(bed_file, data)
     avg_covs = []
-    mean_cov_col = None
     total_len = 0
     with open(sambamba_depth_file) as fh:
-        for line in fh:
-            if line.startswith('#'):
-                mean_cov_col = line.split('\t').index('meanCoverage')
-                continue
-            line_tokens = line.replace('\n', '').split()
+        for line_tokens in (l.rstrip().split() for l in fh if not l.startswith("#")):
+            line_tokens = [x for x in line_tokens if x.strip()]
             start, end = map(int, line_tokens[1:3])
             size = end - start
-            avg_covs.append(float(line_tokens[mean_cov_col]) * size)
+            avg_covs.append(float(line_tokens[-1]) * size)
             total_len += size
     avg_cov = sum(avg_covs) / total_len if total_len > 0 else 0
     return avg_cov
 
-def checkpoint(stem):
-    def check_file(f):
-        def wrapper(*args, **kwargs):
-            out_file = append_stem(args[0], stem)
-            if file_exists(out_file):
-                logger.debug("Skipping %s" % out_file)
-                return out_file
-            return f(*args, **kwargs)
-        return wrapper
-    return check_file
+def _calculate_percentiles(cov_file, dist_file, cutoffs, out_dir, data):
+    """Calculate percentage over over specified cutoff range.
 
-def _calculate_percentiles(in_file, sample, data=None, cutoffs=None):
+    XXX Does not calculate the per-bin coverage estimations which we had
+    earlier with sambamba depth. Instead has a global metric of percent coverage
+    which provides a more defined look at coverage changes by depth.
     """
-    Parse pct bases per region to summarize it in
-    7 different pct of regions points with pct bases covered
-    higher than a completeness cutoff (5, 10, 20, 50 ...)
-    """
-    has_data = False
-    with open(in_file) as in_handle:
-        for i, line in enumerate(in_handle):
-            if i > 0:
-                has_data = True
-                break
-    if not has_data:
+    if not utils.file_exists(dist_file) or not utils.file_exists(cov_file):
         return []
-    out_file = append_stem(in_file, "_summary")
-    out_total_file = append_stem(in_file, "_total_summary")
-    if not utils.file_exists(out_file) or not utils.file_exists(out_total_file):
-        dt = pd.read_csv(in_file, sep="\t", index_col=False)
-        pct = dict()
-        pct_bases = dict()
-        size = np.array(dt["chromEnd"]) - np.array(dt["chromStart"])
-        for cutoff in [h for h in list(dt) if h.startswith("percentage")]:
-            if cutoffs and int(cutoff.split("percentage")[1]) in cutoffs:
-                a = np.array(dt[cutoff])
-                for p_point in [0.01, 10, 25, 50, 75, 90, 99.9]:
-                    q = np.percentile(a, p_point)
-                    pct[(cutoff, p_point)] = q
-                pct_bases[cutoff] = sum(size * a) / float(sum(size))
-
+    sample = dd.get_sample_name(data)
+    out_total_file = append_stem(dist_file, "_total_summary")
+    if not utils.file_exists(out_total_file):
         with file_transaction(data, out_total_file) as tx_file:
             with open(tx_file, 'w') as out_handle:
-                print >>out_handle, "cutoff_reads\tbases_pct\tsample"
-                for k in pct_bases:
-                    print >>out_handle, "\t".join(map(str, [k, pct_bases[k], sample]))
-        with file_transaction(data, out_file) as tx_file:
-            with open(tx_file, 'w') as out_handle:
-                print >>out_handle, "cutoff_reads\tregion_pct\tbases_pct\tsample"
-                for k in pct:
-                    print >>out_handle, "\t".join(map(str, [k[0], k[1], pct[k], sample]))
+                writer = csv.writer(out_handle, dialect="excel-tab")
+                writer.writerow(["cutoff_reads", "bases_pct", "sample"])
+                with open(dist_file) as in_handle:
+                    for line in in_handle:
+                        count, pct = line.strip().split()
+                        count = int(count)
+                        pct = "%.1f" % (float(pct) * 100.0)
+                        if count >= min(cutoffs) and count <= max(cutoffs):
+                            writer.writerow(["percentage%s" % count, pct, sample])
+                    if min(cutoffs) < count:
+                        writer.writerow(["percentage%s" % min(cutoffs), pct, sample])
     # To move metrics to multiqc, will remove older files
     # when bcbreport accepts these one, to avoid errors
     # while porting everything to multiqc
     # These files will be copied to final
-    out_file_fixed = os.path.join(os.path.dirname(out_file), "%s_bcbio_coverage.txt" % sample)
-    out_total_fixed = os.path.join(os.path.dirname(out_file), "%s_bcbio_coverage_avg.txt" % sample)
-    copy_plus(out_file, out_file_fixed)
+    out_total_fixed = os.path.join(os.path.dirname(out_total_file), "%s_bcbio_coverage_avg.txt" % sample)
     copy_plus(out_total_file, out_total_fixed)
-    return [out_file_fixed, out_total_fixed]
+    return [out_total_fixed]
 
-def _read_regions(fn):
+def regions_coverage(bed_file, data):
+    """Generate coverage over regions of interest using mosdepth.
     """
-    Save in a dict the position of regions with
-    the information of the coverage stats.
-    """
-    regions = {}
-    with open(fn) as in_handle:
-        for line in in_handle:
-            if line.startswith("chrom"):
-                regions["header"] = line.strip()
-                continue
-            idx = "".join(line.split("\t")[:2])
-            regions[idx] = line.strip()
-    return regions
+    cov_file, _ = _run_mosdepth(bed_file, data)
+    return cov_file
 
-@checkpoint("_fixed")
-def _add_high_covered_regions(in_file, bed_file, sample, data=None):
+def _run_mosdepth(bed_file, data):
+    """Run mosdepth for a specific BED file generating coverage and distribution.
     """
-    Add regions with higher coverage than the limit
-    as fully covered.
-    """
-    out_file = append_stem(in_file, "_fixed")
-    regions = _read_regions(in_file)
-    with file_transaction(data, out_file) as out_tx:
-        with open(bed_file) as in_handle:
-            with open(out_tx, 'w') as out_handle:
-                if "header" in regions:
-                    print >>out_handle, regions["header"]
-                for line in in_handle:
-                    idx = "".join(line.split("\t")[:2])
-                    if idx not in regions:
-                        print >>out_handle, "%s\t1000\t1000\t100\t100\t100\t100\t100\t100\t100\t100\t100\t100\t%s" % (line.strip(), sample)
-                    else:
-                        print >>out_handle, regions[idx]
-    return out_file
-
-def _summary_variants(in_file, out_file, data=None):
-    """Parse GC and depth variant file
-       to be ready for multiqc.
-    """
-    dt = pd.read_csv(in_file, sep="\t", index_col=False,
-                     dtype={"CG": np.float64, "depth": np.float64}, na_values=["."]).dropna()
-    row = list()
-    with file_transaction(data, out_file) as out_tx:
-        cg = dt["CG"]
-        d = dt["depth"]
-        for p_point in [0.01, 10, 25, 50, 75, 90, 99.9, 100]:
-            if len(cg) > 0:
-                q_cg = np.percentile(cg, p_point)
-            else:
-                q_cg = 0
-            if len(d) > 0:
-                q_d = np.percentile(d, p_point)
-            else:
-                q_d = 0
-            row.append([p_point, q_d, q_cg])
-        pd.DataFrame(row).to_csv(out_tx, header=["pct_variants", "depth", "cg"], index=False, sep="\t")
-
-def regions_coverage(data, bed_file, bam_file, target_name):
-    """Generate coverage over regions of interest using sambamba depth.
-
-    sambamba can segfault with multiple threads so provides a single threaded backup
-    implementation in case of failures.
-    """
+    bam_file = dd.get_align_bam(data) or dd.get_work_bam(data)
     work_dir = utils.safe_makedir(os.path.join(dd.get_work_dir(data), "coverage", dd.get_sample_name(data)))
-    out_file = os.path.join(work_dir, target_name + "_regions_depth.bed")
-    if utils.file_uptodate(out_file, bam_file) and utils.file_uptodate(out_file, bed_file):
-        return out_file
-    with file_transaction(data, out_file) as tx_out_file:
-        try:
-            cmdl = sambamba.make_command(data, "depth region", bam_file, bed_file)
-            cmdl += " -o " + tx_out_file
-            message = "Calculating regions coverage of {target_name} in {bam_file}"
-            do.run(cmdl, message.format(**locals()))
-        except subprocess.CalledProcessError:
-            cmdl = sambamba.make_command(data, "depth region", bam_file, bed_file,
-                                         multicore=False)
-            cmdl += " -o " + tx_out_file
-            message = "Calculating regions coverage of {target_name} in {bam_file} -- single thread backup"
-            do.run(cmdl, message.format(**locals()))
-    return out_file
+    target_name = utils.splitext_plus(os.path.basename(bed_file))[0]
+    out_file = os.path.join(work_dir, "%s-coverage.bed" % target_name)
+    dist_file = os.path.join(work_dir, "%s-distribution.txt" % target_name)
+    if not utils.file_uptodate(out_file, bam_file) or not utils.file_uptodate(out_file, bed_file):
+        with file_transaction(data, out_file) as tx_out_file:
+            with file_transaction(data, dist_file) as tx_dist_file:
+                num_cores = dd.get_cores(data)
+                cmd = ("mosdepth -t {num_cores} -F 1804 -b {bed_file} -d {tx_dist_file} {bam_file} > {tx_out_file}")
+                message = "Calculating regions coverage of {target_name} in {bam_file}"
+                do.run(cmd.format(**locals()), message.format(**locals()))
+    return out_file, dist_file
 
 def coverage_region_detailed_stats(data, out_dir, extra_cutoffs=None):
     """
@@ -338,23 +248,11 @@ def coverage_region_detailed_stats(data, out_dir, extra_cutoffs=None):
     bed_file = dd.get_coverage(data)
     if not bed_file or not utils.file_exists(bed_file):
         return []
-    work_dir = safe_makedir(out_dir)
-    cleaned_bed = clean_file(bed_file, data, prefix="cov-", simple=True)
-
-    cutoffs = {1, 5, 10, 20, 50, 100, 250, 500, 1000, 5000, 10000, 50000}
-
-    with chdir(work_dir):
-        in_bam = dd.get_align_bam(data) or dd.get_work_bam(data)
-        sample = dd.get_sample_name(data)
-        logger.debug("doing coverage for %s" % sample)
-        parse_file = os.path.join(sample + "_coverage.bed")
-        if utils.file_uptodate(parse_file, cleaned_bed) and utils.file_uptodate(parse_file, in_bam):
-            pass
-        else:
-            with file_transaction(data, parse_file) as out_tx:
-                depth_thresholds = sorted(list(cutoffs | extra_cutoffs))
-                cmdl = sambamba.make_command(data, "depth region", in_bam, cleaned_bed, depth_thresholds=depth_thresholds)
-                cmdl += " | sed 's/# chrom/chrom/' > " + out_tx
-                do.run(cmdl, "Run coverage regional analysis for {}".format(sample))
-        out_files = _calculate_percentiles(os.path.abspath(parse_file), sample, data=data, cutoffs=cutoffs)
-    return [os.path.abspath(x) for x in out_files]
+    else:
+        bed_file = clean_file(bed_file, data, prefix="cov-", simple=True)
+        cov_file, dist_file = _run_mosdepth(bed_file, data)
+        cutoffs = {1, 5, 10, 20, 50, 100, 250, 500, 1000, 5000, 10000, 50000}
+        if extra_cutoffs:
+            cutoffs = sorted(list(cutoffs | extra_cutoffs))
+        out_files = _calculate_percentiles(cov_file, dist_file, cutoffs, out_dir, data)
+        return [os.path.abspath(x) for x in out_files]