+++ HISTORY.md
@@ -18,6 +18,8 @@
 - Use samtools multicore indexing, replacing sambamba multicore index.
 - Replace components of pipeline using single core sambamba view -c with
   parallel samtools equivalents.
+- Replace sambamba depth coverage calculations with mosdepth to improve
+  speed and parallelization.
 - Multicore base quality score recalibration with GATK4 and Sentieon.
 - GATK4: add support for gVCF based joint calling.
 - GATK4: fix option usage for gVCF creation with HaplotypeCaller